c renamer dine-opplysninger til dine-opplysninger-ny
