+++ components/forsiden/dine-opplysninger.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 
 jest.mock('next/router', () => require('next-router-mock'));
-import DineOpplysninger from './dine-opplysninger';
+import DineOpplysninger from './dine-opplysninger-ny';
 
 describe('tester komponenten DineOpplysninger', () => {
     test('komponenten rendrer som forventet', () => {
